Move decorator to separate private helper module to prevent circular dependencies
